Trying to fix the IsUserValid test

+++ src/ABC.Testing/MathLibraryTests.cs
@@ -6,8 +6,8 @@ namespace ABC.Testing
         public void CalcSumOfList()
         {
             Calculator.Math math = new Calculator.Math();
-            List<int> numbers = new List<int> { 1, 2, 3, 4, 5 };
-            int result = math.SumOf(numbers);
+            List<double> numbers = new List<double> { 1.0, 2.0, 3.0, 4.0, 5.0 };
+            double result = math.SumOf(numbers);
 
             Assert.NotNull(result);
             Assert.Equal(15, result);
